% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{addGeoCoord}
\alias{addGeoCoord}
\alias{overideGeoCoord}
\title{Add \code{geoCoord} to An Echarts Object}
\usage{
addGeoCoord(chart, geoCoord = NULL, mode = c("add", "overide"))

overideGeoCoord(chart, geoCoord = NULL)
}
\arguments{
\item{chart}{chart \code{echarts} object generated by \code{\link{echart}} or
\code{\link{echartR}}}

\item{geoCoord}{Two mode: \cr
\describe{
 \item{list}{A typical geoCoord list should be: \code{list(place1=c(lng1, lat1), ...)}}
 \item{data.frame}{A data.frame of 3 columns: \code{place}, \code{longitude}, \code{latitude}}
}}

\item{mode}{'add' or 'overide' \code{geoCoord} to the echarts object. Default 'add'.
You can use \code{\link{overideGeoCoord}} to overide existing \code{geoCoord}.}
}
\value{
A modified echarts object
}
\description{
Add \code{geoCoord} object to echarts object. It is used for maps. \cr \cr
For maps with a timeline, \code{geoCoord} object is added to \code{options[[1]]$series[[1]]},
while for those without a timeline, it is added to \code{series[[1]]}.
}
\examples{
\dontrun{

}
}
\references{
\url{http://echarts.baidu.com/echarts2/doc/option.html#title~series-i(map).geoCoord}
}

