% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{addHeatmap}
\alias{addHeatmap}
\alias{overideHeatmap}
\title{Add \code{heatmap} to Echarts Object}
\usage{
addHeatmap(chart, series = NULL, timeslots = NULL, data = NULL,
  gradientColors = list("blue", "cyan", "limegreen", "yellow", "orange",
  "red"), blurSize = 30, minAlpha = 0.05, valueScale = 1, opacity = 1,
  mode = c("add", "overide"), ...)

overideHeatmap(chart, ...)
}
\arguments{
\item{chart}{\code{echarts} object generated by \code{\link{echart}} or
\code{\link{echartR}}.}

\item{series}{Vector. Specify which series you want to insert heatmap. Could be
numeric (index of the series) or string (series name). If NULL, then apply to all.
Default NULL.}

\item{timeslots}{Vector. Specify which timeslots (z) you want to insert heatmap.
Could be numeric (index of the timeslot) or string (timeslot name).
If NULL, then apply to all. Default NULL. You could use \code{z} for short.}

\item{data}{The heatmap source data. Two modes:
\describe{
  \item{data.frame or matrix}{A data.frame or matrix of 3 columns: lng (-180 ~ 180),
  lat (-90 ~ 90) and heat value (0-1). If heat value is out of the range [0,1],
  then it will be normalized.}
  \item{list}{A list in the form: \code{list(c(<lng1>, <lat1>, <value1>),
  c(<lng2>, <lat2>, <value2>), ...)}}
}}

\item{gradientColors}{Color palette of heat visualization. Three mode:
\describe{
  \item{vector}{A vector of colors (hex, name, or rgba string)}
  \item{data.frame or matrix}{A data.frame or matrix of 2 columns: offset (0-1),
  and color (hex, name or rgba string). }
  \item{list}{A list in the form: \code{list(list(offset=0.2, color='red'),
  list(offset=0.5, color='green'), ...)}}
} \cr
Default c('blue', 'cyan', 'lime', 'yellow', 'orange', 'red').}

\item{blurSize}{Size of blur. Default 30.}

\item{minAlpha}{Minimal alpha value which any opacity value below this threshold
will be set \code{minAlpha} in order to prevent over-transparency. Default 0.05.}

\item{valueScale}{Numeric. Scale of value that all the heat value will be multiplied
by this value before plotting. Default 1.}

\item{opacity}{Numeric 0-1. Degree of opacity. Default 1.}

\item{mode}{'add' or 'overide' \code{heatmap} to the echarts object. Default 'add'.
You can use \code{\link{overideHeatmap}} to overide existing \code{heatmap}.}

\item{...}{Elipsis}
}
\value{
A modified echarts object
}
\description{
Add heatmap object to an Echarts object. (applicable for map)
}
\examples{
\dontrun{

}
}
\references{
\url{http://echarts.baidu.com/echarts2/doc/option.html#title~series-i(map).heatmap}
}

