% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{addMarkLine}
\alias{addML}
\alias{addMarkLine}
\alias{addMarkline}
\alias{overideML}
\alias{overideMarkLine}
\alias{overideMarkline}
\title{Add \code{markLine} to An Echarts Object}
\usage{
addMarkLine(chart, series = NULL, timeslots = NULL, data = NULL,
  clickable = TRUE, symbol = c("circle", "arrow"), symbolSize = c(2, 4),
  symbolRotate = NULL, large = FALSE, smooth = FALSE, smoothness = 0.2,
  precision = 2, bundling = list(enable = FALSE, maxTurningAngle = 45),
  effect = list(show = FALSE), itemStyle = NULL, mode = c("add",
  "overide"), ...)

addML(chart, series = NULL, timeslots = NULL, data = NULL,
  clickable = TRUE, symbol = c("circle", "arrow"), symbolSize = c(2, 4),
  symbolRotate = NULL, large = FALSE, smooth = FALSE, smoothness = 0.2,
  precision = 2, bundling = list(enable = FALSE, maxTurningAngle = 45),
  effect = list(show = FALSE), itemStyle = NULL, mode = c("add",
  "overide"), ...)

addMarkline(chart, series = NULL, timeslots = NULL, data = NULL,
  clickable = TRUE, symbol = c("circle", "arrow"), symbolSize = c(2, 4),
  symbolRotate = NULL, large = FALSE, smooth = FALSE, smoothness = 0.2,
  precision = 2, bundling = list(enable = FALSE, maxTurningAngle = 45),
  effect = list(show = FALSE), itemStyle = NULL, mode = c("add",
  "overide"), ...)

overideMarkLine(chart, ...)

overideMarkline(chart, ...)

overideML(chart, ...)
}
\arguments{
\item{chart}{\code{echarts} object generated by \code{\link{echart}} or \code{\link{echartR}}}

\item{series}{Numeric (series index) or character (series name), numeric preferred.
If set NULL, then apply the \code{markPoint} to all the series.}

\item{timeslots}{Numeric (timeslot index) or character (timeslot name), numeric is preferred.
If set NULL, then apply the \code{markPoint} to all the timeslots. You can use \code{z}
for short.}

\item{data}{Data.frame, the data of the \code{markLine}s. It must contain the
following columns: \code{\strong{name1}}, \code{name2, value | type} and/or
\code{x1 | xAxis1, y1 | yAxis1, x2 | xAxis2, y2 | yAxis2} and/or \code{series}. \cr \cr
When series is given, it must be cooresponding to the \code{series} argument.
\describe{
 \item{generic}{generic form is columns \code{[name1, name2, value, x1, y1, x2, y2]}.
 The minimum form is \code{[name1, x1, y1, x2, y2]}}
 \item{scatter, line, bar}{\code{[name1, name2, value, x1, y1, x2, y2]} or
   \code{[name1, type]} (type can only be 'max', 'min' or 'average') or
   \code{[name1, name2, value, xAxis1, yAxis1, xAxis2, yAxis2]}}
 \item{k, eventRiver}{\code{[name1, name2, value, x1, y1, x2, y2]} or
   \code{[name1, name2, value, xAxis1, yAxis1, xAxis2, yAxis2]}}
 \item{map}{\code{[name1, name2, value]}. \strong{You need to pass in \code{
 [name1, lng1, lat1], [name2, lng2, lat2]} using \code{\link{addGeoCoord}}} separately.}
}}

\item{clickable}{Logical, if the graphs are clickable. Default TRUE.}

\item{symbol}{Symbol vector of the markLines, refer to \code{recharts:::validSymbols}.
Default \code{c('circle', 'arrow')}.}

\item{symbolSize}{Numeric vector or \code{JS} function defining the size of the
the beginning symbol and the end symbol of the marklines. Default \code{c(2, 4)}.}

\item{symbolRotate}{Numeric -180 ~ 180. Default NULL.}

\item{large}{Logical if large effect is on. Default FALSE.}

\item{effect}{List. effect configurator of markLines. Default NULL, which is
\code{list(show=FALSE, loop=TRUE, period=15, scaleSize=2, color=NULL,
shadowColor=NULL, shadowBlur=0)}}

\item{itemStyle}{List. It is a list with the structure \code{list(normal=list(...),
emphasis=list(...))}. Default NULL.}

\item{mode}{'add' or 'overide' the data part of \code{markLine} to the echarts
object. Default 'add'.
You can use \code{\link{overideMarkLine}} to overide the data of the existing
\code{markLine}.}

\item{...}{Elipsis}
}
\description{
Add \code{markLine} to An Echarts Object
}
\examples{
\dontrun{
g <- echartr(iris, Sepal.Width, Petal.Width, Species)
g \%>\% addML(c('setosa', 'versicolor'), data=data.frame(name1=rep('mean', 2),
           type=rep('average', 2), series=c('setosa', 'versicolor')))
}

}
\references{
\url{http://echarts.baidu.com/echarts2/doc/option.html#tooltip-line1~series-i(line).markLine}
}

