% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{addMarkPoint}
\alias{addMP}
\alias{addMarkPoint}
\alias{addMarkpoint}
\alias{overideMP}
\alias{overideMarkPoint}
\alias{overideMarkpoint}
\title{Add \code{markPoint} to An Echarts Object}
\usage{
addMarkPoint(chart, series = NULL, timeslots = NULL, data = NULL,
  clickable = TRUE, symbol = "pin", symbolSize = 10,
  symbolRotate = NULL, large = FALSE, effect = list(show = FALSE),
  itemStyle = NULL, ...)

addMP(chart, series = NULL, timeslots = NULL, data = NULL,
  clickable = TRUE, symbol = "pin", symbolSize = 10,
  symbolRotate = NULL, large = FALSE, effect = list(show = FALSE),
  itemStyle = NULL, ...)

addMarkpoint(chart, series = NULL, timeslots = NULL, data = NULL,
  clickable = TRUE, symbol = "pin", symbolSize = 10,
  symbolRotate = NULL, large = FALSE, effect = list(show = FALSE),
  itemStyle = NULL, ...)

overideMarkPoint(chart, ...)

overideMarkpoint(chart, ...)

overideMP(chart, ...)
}
\arguments{
\item{chart}{\code{echarts} object generated by \code{\link{echart}} or \code{\link{echartR}}}

\item{series}{Numeric (series index) or character (series name), numeric preferred.
If set NULL, then apply the \code{markPoint} to all the series.}

\item{timeslots}{Numeric (timeslot index) or character (timeslot name), numeric preferred.
If set NULL, then apply the \code{markPoint} to all the timeslots. You can use \code{z}
for short.}

\item{data}{Data.frame, the data of the \code{markPoint}s. It must contain the
following columns: \code{\strong{name}}, and/or \code{value | type} and/or
\code{x | xAxis, y | yAxis} and/or \code{series}. \cr \cr
When series is given, it must be corresponding to the \code{series} argument.
\describe{
 \item{generic}{generic form is columns \code{[name, x, y]} or \code{[name, value,
   x, y]}}
 \item{scatter, line, bar}{\code{[name, value, x, y]} or \code{[name, type]}
 (type can only be 'max' or 'min') or \code{[name, value, xAxis, yAxis]}}
 \item{k, eventRiver}{\code{[name, value, x, y]} or \code{[name, value, xAxis, yAxis]}}
 \item{map}{\code{[name, value]}. \strong{You need to pass in \code{[name, lng, lat]}
 using \code{\link{addGeoCoord}}} separately.}
}
Note that markLine dataset is compatible with \code{addMarkPoint}.
\itemize{
 \item \code{x1, x2} are treated as \code{x} (keep the first detected one only)
 \item \code{y1, y2} are treated as \code{y} (keep the first detected one only)
 \item \code{xAxis1, xAxis2} are treated as \code{xAxis} (keep the first detected
  one only)
 \item \code{yAxis1, yAxis2} are treated as \code{yAxis} (keep the first detected
  one only
}}

\item{clickable}{Logical, if the points are clickable. Default TRUE.}

\item{symbol}{Symbol of the markpoints, refer to \code{recharts:::validSymbols}.
Default 'pin'.}

\item{symbolSize}{Numeric or vector \code{c(height, width)} or JS function. Default 10.}

\item{symbolRotate}{Numeric -180 ~ 180. Default NULL.}

\item{large}{Logical if large effect is on. Default FALSE.}

\item{effect}{List. effect configurator of markPoints. Default NULL, which is
\code{list(show=FALSE, type='scale', loop=TRUE, period=15, scaleSize=2,
bounceDistance=10, color=NULL, shadowColor=NULL, shadowBlur=0)}}

\item{itemStyle}{List. It is a list with the structure \code{list(normal=list(...),
emphasis=list(...))}. Default NULL.}

\item{...}{Elipsis}

\item{mode}{'add' or 'overide' the data part of \code{markPoint} to the echarts
object. Default 'add'.
You can use \code{\link{overideMarkPoint}} to overide the data of the existing
\code{markPoint}.}
}
\description{
Add \code{markPoint} to An Echarts Object
}
\examples{
\dontrun{
g <- echartr(iris, Sepal.Width, Petal.Width, Species)
g \%>\% addMP(c('setosa', 'versicolor'), data=data.frame(name=c('max', 'min'),
           type=c('max', 'min'), series=c('setosa', 'versicolor')))
}

}
\references{
\url{http://echarts.baidu.com/echarts2/doc/option.html#tooltip-line1~series-i(line).markPoint}
}
\seealso{
\code{\link{addMarkLine}}  \code{\link{addMarkline}}
}

