% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{addNameMap}
\alias{addNameMap}
\alias{overideNameMap}
\title{Add \code{nameMap} to Echarts Object}
\usage{
addNameMap(chart, nameMap, mode = c("add", "overide"))

overideNameMap(chart, ...)
}
\arguments{
\item{chart}{\code{echarts} object generated by \code{\link{echart}} or
\code{\link{echartR}}.}

\item{nameMap}{Two modes: \cr
\describe{
 \item{data.frame/matrix}{A data.frame or matrix comprising of 2 columns: \code{nameToTranslate &
 nameTranslatedTo}. E.g., You can load a preinstalled Chinese-English geographic
 dictionary using \code{recharts:::geoNameMap}.}
 \item{list}{The nameMap in list should follow the structure: \code{list(list(
`United States of America`='USA'), list('United Kingdom'='GB'), ...)}}
}}

\item{mode}{'add' or 'overide' \code{nameMap} to the echarts object. Default 'add'.
You can use \code{\link{overideNameMap}} to overide existing \code{nameMap}.}
}
\value{
A modified echarts object
}
\description{
For map charts, you can add \code{nameMap} to translate the place names from one to another.
The \code{nameMap} object will be inserted to the first series of the echarts object.
}
\examples{
\dontrun{
}
}
\references{
\url{http://echarts.baidu.com/echarts2/doc/option.html#title~series-i(map).nameMap}
}
\seealso{
\code{data(geoNameMap)}
}

