% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoCoord.R
\name{convBD2GCJ}
\alias{convBD2GCJ}
\title{Transform BD-09 (Baidu) coordinates to GCJ-02 (Google CN/Amap)}
\usage{
convBD2GCJ(y, ...)
}
\arguments{
\item{y}{\itemize{
\item A vector \code{c(latitude, longitude)} \cr
\item simply latitude \cr
\item a matrix (row 1-2 or col 1-2). The function will choose how to read the data \cr
\item a data.frame (col 1-2)
}}

\item{...}{\itemize{
\item When \code{y} is only latitude, you can pass in \code{x} (longitude) here. \cr
\item when \code{y} is a vector of \code{c(lat, lon)}, you can pass in the rest vectors as well.
\item when \code{y} is a matrix or a data.frame, ... is omitted.
}}
}
\value{
A 2-col data.frame ([lng, lat]) of transformed coordinates.
}
\description{
GCJ-02 is coordinate system mandated by Chinese Gov't. BD-09 is Baidu specific
coordinates that encrypts GCJ-02 further more.
This function decrypts BD-09 back to GCJ-02 coordinates (Chinese coordinates).
}
\note{
Latitude is the horizontal line serving as y-axis metric, longitude is
the vertical line serving as x-axis metric.
}
\section{GIS}{

}
\examples{
\dontrun{
# Tiananmen square's BD-06 coordinate is c(39.91509, 116.40350)
# http://www.google.cn/maps/place/Tiananmen,+Dongcheng,+Beijing/@39.90875,116.39713,16z?hl=en

## Single point
convBD2GCJ(c(39.91509, 116.40350))  # or
convBD2GCJ(39.91509, 116.40350)  # get
#           lat      lng
# [1,] 39.90875 116.3971

## Multiple points
### Vectors
convBD2GCJ(c(39.91509, 116.40350), c(39.91509, 116.40350))  # get
#           lat      lng
# [1,] 39.90875 116.3971
# [2,] 39.90875 116.3971

### Matrix
m <- matrix(c(39.91509, 116.40350, 39.91509, 116.40350, 39.91509, 116.40350), nrow=2)
m
#           [,1]      [,2]      [,3]
# [1,]  39.91509  39.91509  39.91509
# [2,] 116.40350 116.40350 116.40350
convBD2GCJ(m)  # get
#           lat      lng
# [1,] 39.90875 116.3971
# [2,] 39.90875 116.3971
# [3,] 39.90875 116.3971

### data.frame
df <- data.frame(lat=c(39.91509, 39.91509, 39.91509, NA),
                 lon=c(116.40350, 116.40350, 116.40350, 116.40350))
convBD2GCJ(df)  # get
#           lat      lng
# [1,] 39.90875 116.3971
# [2,] 39.90875 116.3971
# [3,] 39.90875 116.3971
# [4,]       NA       NA
}

}
\author{
Yiying Wang, \email{wangy@aetna.com}
}
\references{
\url{https://on4wp7.codeplex.com/SourceControl/changeset/view/21483#353936}
}
\seealso{
\code{\link{convBD2GCJ}}, \code{\link{convCoord}}.
}

