% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoCoord.R
\name{convBD2WGS}
\alias{convBD2WGS}
\title{Transform BD-09 (Baidu) coordinates to WGS-84 (Global Coord)}
\usage{
convBD2WGS(y, ...)
}
\arguments{
\item{y}{\itemize{
\item A vector \code{c(latitude, longitude)} \cr
\item simply latitude \cr
\item a matrix (row 1-2 or col 1-2). The function will choose how to read the data \cr
\item a data.frame (col 1-2)
}}

\item{...}{\itemize{
\item When \code{y} is only latitude, you can pass in \code{x} (longitude) here. \cr
\item when \code{y} is a vector of \code{c(lat, lon)}, you can pass in the rest vectors as well.
\item when \code{y} is a matrix or a data.frame, ... is omitted.
}}
}
\value{
A 2-col data.frame ([lng, lat]) of transformed coordinates.
}
\description{
BD-09 is Baidu specific coordinates that encrypts GCJ-02 further more.
This function decrypts BD-09 back to WGS-84 coordinates (Global coordinates).
}
\note{
Latitude is the horizontal line serving as y-axis metric, longitude is
the vertical line serving as x-axis metric.
}
\section{GIS}{

}
\examples{
\dontrun{
# Tiananmen square's BD-06 coordinate is c(39.91509, 116.40350)
# http://www.openstreetmap.org/#map=19/39.90734/116.39089

## Single point
convBD2WGS(c(39.91509, 116.40350))  # or
convBD2WGS(39.91509, 116.40350)  # get
#           lat      lng
# [1,] 39.90734 116.3909

## Multiple points
### Vectors
convBD2WGS(c(39.91509, 116.40350), c(39.91509, 116.40350))  # get
#           lat      lng
# [1,] 39.90734 116.3909
# [2,] 39.90734 116.3909

### Matrix
m <- matrix(c(39.91509, 116.40350, 39.91509, 116.40350, 39.91509, 116.40350), nrow=2)
m
#           [,1]      [,2]      [,3]
# [1,]  39.90734  39.90734  39.90734
# [2,] 116.39089 116.39089 116.39089
convBD2WGS(m)  # get
#           lat      lng
# [1,] 39.90734 116.3909
# [2,] 39.90734 116.3909
# [3,] 39.90734 116.3909

### data.frame
df <- data.frame(lat=c(39.91509, 39.91509, 39.91509, NA),
                 lon=c(116.40350, 116.40350, 116.40350, 116.40350))
convBD2WGS(df)  # get
#           lat      lng
# [1,] 39.90734 116.3909
# [2,] 39.90734 116.3909
# [3,] 39.90734 116.3909
# [4,]       NA       NA
}

}
\author{
Yiying Wang, \email{wangy@aetna.com}
}
\references{
\url{https://on4wp7.codeplex.com/SourceControl/changeset/view/21483#353936}
}
\seealso{
\code{\link{convWGS2BD}}, \code{\link{convBD2GCJ}}, \code{\link{convCoord}}.
}

