% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoCoord.R
\name{convGCJ2WGS}
\alias{convGCJ2WGS}
\title{Transform GCJ-02 (Google CN/Amap) coordinates to WGS-84 (Global)}
\usage{
convGCJ2WGS(y, ...)
}
\arguments{
\item{y}{\itemize{
\item A vector \code{c(latitude, longitude)} \cr
\item simply latitude \cr
\item a matrix (row 1-2 or col 1-2). The function will choose how to read the data \cr
\item a data.frame (col 1-2)
}}

\item{...}{\itemize{
\item When \code{y} is only latitude, you can pass in \code{x} (longitude) here. \cr
\item when \code{y} is a vector of \code{c(lat, lon)}, you can pass in the rest vectors as well.
\item when \code{y} is a matrix or a data.frame, ... is omitted.
}}
}
\value{
A 2-col data.frame ([lng, lat]) of transformed coordinates.
}
\description{
GCJ-02 is coordinate system mandated by Chinese Gov't.
This function decrypts it back to WGS-84 coordinates (gloabal coordinates).
}
\note{
Latitude is the horizontal line serving as y-axis metric, longitude is
the vertical line serving as x-axis metric.
}
\section{GIS}{

}
\examples{
\dontrun{
# Tiananmen square's GCJ-02 coordinate is c(39.908746, 116.397131)
# http://www.openstreetmap.org/#map=19/39.90734/116.39089

## Single point
convGCJ2WGS(c(39.908746, 116.397131))  # or
convGCJ2WGS(39.908746, 116.397131)  # get
#           lat      lng
# [1,] 39.90734 116.3909

## Multiple points
### Vectors
convGCJ2WGS(c(39.908746, 116.397131), c(39.908746, 116.397131))  # get
#           lat      lng
# [1,] 39.90734 116.3909
# [2,] 39.90734 116.3909

### Matrix
m <- matrix(c(39.908746, 116.397131, 39.908746, 116.397131, 39.908746, 116.397131), nrow=2)
m
#           [,1]      [,2]      [,3]
# [1,]  39.90875  39.90875  39.90875
# [2,] 116.39713 116.39713 116.39713
convGCJ2WGS(m)  # get
#           lat      lng
# [1,] 39.90734 116.3909
# [2,] 39.90734 116.3909
# [3,] 39.90734 116.3909

### data.frame
df <- data.frame(lat=c(39.908746, 39.908746, 39.908746, NA),
                 lon=c(116.397131, 116.397131, 116.397131, 116.397131))
convGCJ2WGS(df)  # get
#           lat      lng
# [1,] 39.90734 116.3909
# [2,] 39.90734 116.3909
# [3,] 39.90734 116.3909
# [4,]       NA       NA
}

}
\author{
Yiying Wang, \email{wangy@aetna.com}
}
\references{
\url{https://on4wp7.codeplex.com/SourceControl/changeset/view/21483#353936}
}
\seealso{
\code{\link{convWGS2GCJ}}, \code{\link{convCoord}}.
}

