% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoCoord.R
\name{convWGS2BD}
\alias{convWGS2BD}
\title{Transform WGS-84 (Global GeoSys) coordinates to BD-09 (Baidu)}
\usage{
convWGS2BD(y, ...)
}
\arguments{
\item{y}{\itemize{
\item A vector \code{c(latitude, longitude)} \cr
\item simply latitude \cr
\item a matrix (row 1-2 or col 1-2). The function will choose how to read the data \cr
\item a data.frame (col 1-2)
}}

\item{...}{\itemize{
\item When \code{y} is only latitude, you can pass in \code{x} (longitude) here. \cr
\item when \code{y} is a vector of \code{c(lat, lon)}, you can pass in the rest vectors as well.
\item when \code{y} is a matrix or a data.frame, ... is omitted.
}}
}
\value{
A 2-col data.frame ([lng, lat]) of transformed coordinates.
}
\description{
BD-09 is Baidu specific coordinates that encrypts GCJ-02 further more.
This function encrypts WGS-84 (Global) into BD-09 coordinates (Baidu coordinates).
}
\note{
Latitude is the horizontal line serving as y-axis metric, longitude is
the vertical line serving as x-axis metric.
}
\section{GIS}{

}
\examples{
\dontrun{
# Tiananmen square's WGS-84 coordinate is c(39.90734, 116.39089)
# http://api.map.baidu.com/marker?location=39.91509,116.40350&title=Tiananmen&content=Tiananmen\%20square&output=html

## Single point
convWGS2BD(c(39.90734, 116.39089))  # or
convWGS2BD(39.90734, 116.39089)  # get
#           lat      lng
# [1,] 39.91508 116.4035

## Multiple points
### Vectors
convWGS2BD(c(39.90734, 116.39089), c(39.90734, 116.39089))  # get
#           lat      lng
# [1,] 39.91508 116.4035
# [2,] 39.91508 116.4035

### Matrix
m <- matrix(c(39.90734, 116.39089, 39.90734, 116.39089, 39.90734, 116.39089), nrow=2)
m
#           [,1]      [,2]      [,3]
# [1,]  39.90734  39.90734  39.90734
# [2,] 116.39089 116.39089 116.39089
convWGS2BD(m)  # get
#           lat      lng
# [1,] 39.91508 116.4035
# [2,] 39.91508 116.4035
# [3,] 39.91508 116.4035

### data.frame
df <- data.frame(lat=c(39.90734, 39.90734, 39.90734, NA),
                 lon=c(116.39089, 116.39089, 116.39089, 116.39089))
convWGS2BD(df)  # get
#           lat      lng
# [1,] 39.91508 116.4035
# [2,] 39.91508 116.4035
# [3,] 39.91508 116.4035
# [4,]       NA       NA
}

}
\author{
Yiying Wang, \email{wangy@aetna.com}
}
\references{
\url{https://on4wp7.codeplex.com/SourceControl/changeset/view/21483#353936}
}
\seealso{
\code{\link{convBD2WGS}}, \code{\link{convWGS2GCJ}}, \code{\link{convCoord}}
}

