% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoCoord.R
\name{convWGS2GCJ}
\alias{convWGS2GCJ}
\title{Transform WGS-84 (Global) coordinates to GCJ-02 (Google CN/Amap)}
\usage{
convWGS2GCJ(y, ...)
}
\arguments{
\item{y}{\itemize{
\item A vector \code{c(latitude, longitude)} \cr
\item simply latitude number \cr
\item a matrix (row 1-2 or col 1-2). The function will choose how to read the data \cr
\item a data.frame (col 1-2)
}}

\item{...}{\itemize{
\item When \code{y} is only latitude, you can pass in \code{x} (longitude) here. \cr
\item when \code{y} is a vector of \code{c(lat, lon)}, you can pass in the rest vectors as well.
\item when \code{y} is a matrix or a data.frame, ... is omitted.
}}
}
\value{
A 2-col data.frame ([lng, lat]) of transformed coordinates.
}
\description{
WGS-84 coordinates are gloabal coordinates.
This function encrypts it into GCJ-02 that is mandated by Chinese Gov't.
}
\note{
Latitude is the horizontal line serving as y-axis metric, longitude is
the vertical line serving as x-axis metric.
}
\section{GIS}{

}
\examples{
\dontrun{
# Tiananmen square's WGS-84 coordinate is c(39.90734, 116.39089)
# http://www.google.cn/maps/place/Tiananmen,+Dongcheng,+Beijing/@39.90874,116.39713,16z?hl=en

## Single point
convWGS2GCJ(c(39.90734, 116.39089))  # or
convWGS2GCJ(39.90734, 116.39089)  # get
#           lat      lng
# [1,] 39.90874 116.3971

## Multiple points
### Vectors
convWGS2GCJ(c(39.90734, 116.39089), c(39.90734, 116.39089))  # get
#           lat      lng
# [1,] 39.90874 116.3971
# [2,] 39.90874 116.3971

### Matrix
m <- matrix(c(39.90734, 116.39089, 39.90734, 116.39089, 39.90734, 116.39089), nrow=2)
m
#           [,1]      [,2]      [,3]
# [1,]  39.90734  39.90734  39.90734
# [2,] 116.39089 116.39089 116.39089
convWGS2GCJ(m)  # get
#           lat      lng
# [1,] 39.90874 116.3971
# [2,] 39.90874 116.3971
# [3,] 39.90874 116.3971

### data.frame
df <- data.frame(lat=c(39.90734, 39.90734, 39.90734, NA),
                 lon=c(116.39089, 116.39089, 116.39089, 116.39089))
convWGS2GCJ(df)  # get
#           lat      lng
# [1,] 39.90874 116.3971
# [2,] 39.90874 116.3971
# [3,] 39.90874 116.3971
# [4,]       NA       NA
}

}
\author{
Yiying Wang, \email{wangy@aetna.com}
}
\references{
\url{https://on4wp7.codeplex.com/SourceControl/changeset/view/21483#353936}
}
\seealso{
\code{\link{convGCJ2WGS}}, \code{\link{convCoord}}.
}

