% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echart.R
\name{echartr}
\alias{echartR}
\alias{echartr}
\title{Framework Function to Create An Echarts Object}
\usage{
echartr(data, x = NULL, y = NULL, series = NULL, weight = NULL,
  z = NULL, lat = NULL, lng = NULL, type = "auto", subtype = NULL,
  ...)

echartR(data, x = NULL, y = NULL, series = NULL, weight = NULL,
  z = NULL, lat = NULL, lng = NULL, type = "auto", subtype = NULL,
  ...)
}
\arguments{
\item{data}{data.frame.}

\item{x}{independent variable(s). Some charts calls for one, some calls for two.}

\item{y}{dependent variable(s). Most charts calls for one, but some calls for more.}

\item{series}{data series variables.
\describe{
  \item{single coordinate system}{In Cartesian Coordinate System (bar, scatter, k, ...) and
  some other charts (venn), series represents data series. Different data series
  are displayed in the same coordinate system}
  \item{multiple coordinate system}{In polar coordinate system (pie, radar, ...) and
  some other charts (map, tree, ...), series represents index of coordinate system.
  Different coordinate system indices are displayed in paralell coordinate systems.}
}}

\item{weight}{In some charts (bar, bubble, line, ...), weight represents the size
of the graph elements.}

\item{z}{timeline variable. When \code{z} is defined, recharts builds a timeline
widget to show the changes along with time.}

\item{lat}{latitude variable (-180 ~ 180) for map/heatmap}

\item{lng}{longitude variable (-90 ~ 90) for map/heatmap}

\item{type}{chart type. Now recharts supports major types of
\itemize{
  \item \href{http://madlogos.github.io/recharts/Basic_Plots_01_Scatterplot.html}{scatter/point/bubble}
  \item \href{http://madlogos.github.io/recharts/Basic_Plots_02_Bar.html}{bar/column/histogram}
  \item \href{http://madlogos.github.io/recharts/Basic_Plots_03_Line.html}{line/area/curve/wave}
  \item \href{http://madlogos.github.io/recharts/Basic_Plots_04_K.html}{candlestick/k}
  \item \href{http://madlogos.github.io/recharts/Basic_Plots_05_eventRiver.html}{eventRiver}
  \item \href{http://madlogos.github.io/recharts/Basic_Plots_21_Pie.html}{pie/ring/rose}
  \item \href{http://madlogos.github.io/recharts/Basic_Plots_22_Funnel.html}{funnel/pyramid}
  \item \href{http://madlogos.github.io/recharts/Basic_Plots_23_Radar.html}{radar/spider/star}
  \item \href{http://madlogos.github.io/recharts/Basic_Plots_11_Force.html}{force/force_line}
  \item \href{http://madlogos.github.io/recharts/Basic_Plots_12_Chord.html}{chord}
  \item \href{http://madlogos.github.io/recharts/Basic_Plots_24_Gauge.html}{gauge}
  \item \href{http://madlogos.github.io/recharts/Basic_Plots_13_WordCloud.html}{wordCloud}
  \item \href{http://madlogos.github.io/recharts/Basic_Plots_14_Venn.html}{venn}
  \item \href{http://madlogos.github.io/recharts/Basic_Plots_31_Map.html}{map_world/map_china/map_world_multi/map_china_multi}
  \item \href{http://madlogos.github.io/recharts/Basic_Plots_32_Tree.html}{tree/htree/tree_inv/htree_inv}
  \item \href{http://madlogos.github.io/recharts/Basic_Plots_33_Treemap.html}{treemap}
  \item \href{http://madlogos.github.io/recharts/Basic_Plots_15_Heatmap.html}{heatmap}
}}

\item{subtype}{some chart types support subtypes.
\describe{
  \item{bar/column}{"stack"}
  \item{histogram}{c("count","freq","density")}
  \item{line/area}{c("stack","smooth","dotted","solid","dashed")}
  \item{curve/wave}{c("stack","solid")}
  \item{map_world/world_map/map_world_multi/world_map_multi}{c("sum","average",
  [Valid Country Names])}
  \item{map_china/china_map/map_china_multi/china_map_multi}{c("sum","average",
  [Valid China Province Names])}
  \item{pie}{c("multi","clock","clockwise")}
  \item{ring}{c("info","multi","clock","clockwise")}
  \item{rose}{c("area","radius","multi","clock","clockwise")}
  \item{chord}{c("ribbon","scale","scaletext","rotatelab","hidelab","clock",
  "clockwise","asc","desc","ascsub","descsub")}
  \item{force/force_line}{c("arrow","triangle")}
  \item{funnel/pyramid}{c("left","center","right")}
  \item{tree/vtree/htree/tree_inv/vtree_inv/htree_inv}{c("curve","broken",
  "dotted","solid","dashed")}
  \item{radar/spider/star}{"fill"}
}}

\item{...}{}
}
\description{
This function is used to produce the skeleton of an Echarts object with default settings.
You can then tune the settings using <setEchartWidgets> functions, <aesthetic> functions,
and others.
}
\references{
Online Manual: \url{http://madlogos.github.io/recharts}
}

