% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getColFromPal}
\alias{getColFromPal}
\title{Get The Colors Vector From A Named Palette}
\usage{
getColFromPal(palname = NULL, n = 6)
}
\arguments{
\item{palname}{name of the palette. Default NULL to get echarts default. Could be:
\itemize{
 \item \link{RColorBrewer} palettes: Including \code{'BrBG', 'PiYG', 'PRGn', 'PuOr', 'RdBu',
 'RdGy', 'RdYlBu', 'RdYlGn', 'Spectral', 'Accent', 'Dark2', 'Paired', 'Pastel1',
 'Pastel2', 'Set1', 'Set2', 'Set3', 'Blues', 'BuGn', 'BuPu', 'GnBu', 'Greens',
 'Greys', 'Oranges', 'OrRd', 'PuBu', 'PuBuGn', 'PuRd', 'Purples', 'RdPu', 'Reds',
 'YlGn', 'YlGnBu', 'YlOrBr', 'YlOrRd'} \cr
 \item \link{ggthemes} palettes: \code{'calc', 'economist', 'economist_white', 'economist_stata',
 'excel', 'exel_fill', 'excel_line', 'excel_new', 'few', 'fivethirtyeight', '538', 'manyeyes',
 'gdocs', 'pander', 'tableau', 'stata', 'stata1', 'stata1r', 'statamono', 'ptol',
 'tableau20', 'tableau10medium', 'tableaugray', 'tableauprgy', 'tableaublrd',
 'tableaugnor', 'tableaucyclic', 'tableau10light', 'tableaublrd12', 'tableauprgy12',
 'tableaugnor12', 'hc', 'darkunica', 'solarized', 'solarized_red', 'solarized_yellow',
 'solarized_orange', 'solarized_magenta', 'solarized_violet', 'solarized_blue',
 'solarized_cyan', 'solarized_green', 'wsj', 'wsj_rgby', 'wsj_red_green',
 'wsj_black_green', 'wsj_dem_rep', 'colorblind', 'trafficlight'} \cr
 \item Aetna official palettes: Including \code{'aetnagreen', 'aetnablue', 'aetnaviolet',
 'aetnaorange', 'aetnateal', 'aetnacranberry'} \cr
 \item Other palettes: \code{'rainbow', 'terrain', 'topo', 'heat', 'cm'}
}}

\item{n}{length of the color vector when the palette is continuous (\code{rain, cm,
terrain, topo, heat, ...}). Default 6.}
}
\value{
color vectors
}
\description{
Get hex color vector of a named palette from \code{\link{RColorBrewer}}, \code{\link{ggthemes}}
or \code{\link{grDevices}}. You can \code{\link{show_col}} the vector to
see the effects.
}
\examples{
\dontrun{
library(scales)
show_col(getColFromPal('tableau20'))
show_col(getColFromPal('hc'))
}
}
\seealso{
\code{\link{RColorBrewer}}, \code{\link{scales}}, \code{\link{ggthemes}},
\code{\link{show_col}}
}

