% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getColors}
\alias{getColors}
\title{Get Hex Color Vector (Not Exported)}
\usage{
getColors(palette, ...)
}
\arguments{
\item{palette}{Palette, default NULL. Could be
\itemize{
 \item palette name, e.g, "Blues". The palette will be proceeded by \code{\link{getColFromPal}}.
 \item a hex color, e.g., "#FFFFFF", "0xFFFFFFFF"
 \item a vector or color names, hex colors
 \item NULL
}}

\item{...}{Elipsis}
}
\value{
A vector of hex colors
}
\description{
Get color vector from a palette/color name. It is wider than \code{\link{getColFromPal}}.
}
\examples{
\dontrun{
library(scales)
show_col(getColors(NULL))
show_col(getColors("terrain"))
show_col(getColors(c('red', 'gold', 'skyblue')))
}
}
\seealso{
\code{\link{getColFromPal}} \code{\link{RColorBrewer}} \code{\link{ggthemes}}
}

