% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{iif}
\alias{ifblank}
\alias{ifempty}
\alias{ifmissing}
\alias{ifna}
\alias{ifnan}
\alias{ifnull}
\alias{ifzero}
\alias{iif}
\title{If-else Replacement Function For Vectors}
\usage{
iif(x, y, cond = c("is.null", "is.na", "is.nan", "is.blank", "is.zero",
  "missing"))

ifnull(x, y)

ifna(x, y)

ifnan(x, y)

ifblank(x, y)

ifempty(x, y)

ifzero(x, y)

ifmissing(x, y)
}
\arguments{
\item{x}{A vector/list to replace}

\item{y}{The vector to be replaced with}

\item{cond}{Condition string, could be 'is.null', 'is.na', 'is.nan', 'is.blank',
'is.zero', or 'missing'}
}
\description{
If-else Replacement Function For Vectors
}
\examples{
\dontrun{
ifna(c(NA, 1, 4, NA), 0)
# get c(0, 1, 4, 0)
}

}

