% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{reheadHTMLTable}
\alias{reheadHTMLTable}
\title{Reformat HTML Table}
\usage{
reheadHTMLTable(dataset, heading, footRows = 0, align = c("left",
  rep("center", ncol - 1)), concatCol = NULL, caption = NULL,
  tableWidth = "100\%")
}
\arguments{
\item{dataset}{The dataset to draw table.}

\item{heading}{The heading you want to input.
'|' indicates colspan, '=' indicates rowspan.}

\item{footRows}{The last several rows as <tfoot>.}

\item{align}{Alignment of columns.}

\item{concatCol}{Index of columns to concatenate,
to make the table look hierachical.}

\item{caption}{Table caption.}

\item{tableWidth}{Width of the table.}
}
\value{
A table in HTML.
}
\description{
Convert a data frame to an HTML table object and reformat it.
}
\examples{
\dontrun{
## A reformatted table with colspan/colrow=2
heading <- matrix(c("Sepal", NA, "Petal", NA, "Species", "Length", "Width",
                    "Length", "Width", NA), byrow=TRUE, nrow=2)
reheadHTMLTable(head(iris), heading)
}
}
\seealso{
\code{\link{knitr::kable}}
}

