% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{relocWidget}
\alias{relocWidget}
\title{Re-locate Echarts Widgets (Position of Upper-left/Lower-right Point)}
\usage{
relocWidget(chart, widgets, x = NULL, y = NULL, x2 = NULL, y2 = NULL)
}
\arguments{
\item{chart}{Echarts object}

\item{widgets}{Vector or list, could be \code{'title', 'timeline', 'legend', 'toolbox',
'dataRange', 'dataZoom', 'roamController'}}

\item{x}{Vector, x-coordinates of the widgets' upper-left point}

\item{y}{Vector, y-coordinates of the widgets' upper-left point}

\item{x2}{Vector, x-coordinates of the widgets' lower-right point}

\item{y2}{Vector, y-coordinates of the widgets' lower-right point}
}
\value{
A modified Echarts object
}
\description{
Re-locate Echarts Widgets (Position of Upper-left/Lower-right Point)
}
\note{
If \code{x, y, x2, y2} are shorter in length than the list \code{widgets},
the last element of \code{x, y, x2, y2} will be applied to cover the rest.
If \code{x, y, x2, y2} are longer in length than the list \code{widgets},
the redundent elements will be dropped.
}
\examples{
\dontrun{
g <- echartR(iris, Sepal.Width, Petal.Width) \%>\% setDataZoom()
g \%>\% relocWidgets('dataZoom', x=150)
}
}
\seealso{
\code{\link{setGrid}}
}

