% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rgba}
\alias{rgba}
\title{Get A String Containing 'rgba' Function}
\usage{
rgba(vecrgb, ...)
}
\arguments{
\item{vecrgb}{A vector of RGB elements, or simply red int.}

\item{...}{If vecrgb is simply red int, you can pass green, blue, alpha int here.}
}
\value{
A character string. E.g, 'rgba(125, 125, 125, 0.6)' or '#FFFFFF'
}
\description{
Echarts uses rgba function heavily. You can convert color vectors into rgba function
in string form to pass to an echarts object.
}
\examples{
\dontrun{
rgba(c(123, 123, 124, 125))  # return 'rgba(123,123,124,0.490196078431373)'
rgba(123, 123, 124, 0.5) # return 'rgba(123,123,124,0.5)'
rgba(123, 123, 124)  # return '#7B7B7C'
}
}

