% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{setAxis}
\alias{setAxis}
\alias{setX1Axis}
\alias{setXAxis}
\alias{setY1Axis}
\alias{setYAxis}
\title{Set \code{x|yAxis} of Echarts (Primary or Secondary)}
\usage{
setAxis(chart, series = NULL, which = c("x", "y", "x1", "y1"),
  type = c("value", "category", "time", "log"), show = TRUE,
  position = c("bottom", "top", "left", "right"), name = "",
  nameLocation = c("end", "start"), nameTextStyle = emptyList(),
  boundaryGap = c(0, 0), min = NULL, max = NULL, scale = TRUE,
  splitNumber = NULL, axisLine = list(show = TRUE, onZero = FALSE),
  axisTick = list(show = FALSE), axisLabel = list(show = TRUE),
  splitLine = list(show = TRUE), splitArea = list(show = FALSE),
  data = list())

setYAxis(chart, ...)

setY1Axis(chart, ...)

setXAxis(chart, ...)

setX1Axis(chart, ...)
}
\arguments{
\item{chart}{\code{echarts} object generated by \code{\link{echart}} or \code{\link{echartR}}}

\item{series}{Which series to be put on this axis. Could be:
\itemize{
 \item series names, in vectors. E.g, \code{c('setosa', 'virginica')};
 \item index of series, in vectors. E.g., \code{1:2} or \code{c(1,3)}.
}}

\item{which}{Which axis to be modified. Could be one of the following:
\describe{
 \item{x}{primary x axis}
 \item{y}{primary y axis}
 \item{x1}{secondary x axis}
 \item{y1}{secondary y axis}
}}

\item{type}{Type of the axis. Could be \code{c('time', 'value', 'category', 'log')}.
Default 'value'.}

\item{show}{Logical, whether to show this axis. Default TRUE.}

\item{position}{Position of this axis. Could be \code{c('bottom', 'top', 'left',
'right')} (default for primary x axis, secondary x axis, primary y axis and secondary
y axis, respectively.)}

\item{name}{Name of this axis. Default not set ("").}

\item{nameLocation}{Axis name location. Could be \code{c('end', 'start')}. Default
'end'.}

\item{nameTextStyle}{Axis name text style. Could be a list of \code{textStyle}
features. Default following textStyle global settings.}

\item{boundaryGap}{A two-element numeric vector, defining the policy of the space
at the two ends of the axis (percents). Deafult \code{c(0, 0)}.}

\item{min}{The mininum value of the axis. Default NULL (automatic). If a numeric
value is set, \code{boundaryGap} is disabled.}

\item{max}{The maxinum value of the axis. Default NULL (automatic). If a numeric
value is set, \code{boundaryGap} is disabled.}

\item{scale}{Logical, for axis of 'value', 'time', 'log' type, to define whether
zoom the scale to the range between _min and _max. Default TRUE.}

\item{splitNumber}{Numeric, how many sections to devide the axis. Default NULL,
automatically deviding based on algorithms of \code{min} and \code{max}.}

\item{axisLine}{A list. Default: \cr
\code{list(show=TRUE, onZero=FALSE, lineStyle=list( \cr
type='solid', color='#48b', width=2, shadowColor='rgba(0,0,0,0)', shadowBlur=5,
shadowOffsetX=3, shadowOffsetY=3))} \cr \cr
\code{lineStyle} accepts features \code{color, width, type, shadowColor, shadowBlur,
shadowOffsetX, shadowOffsetY}}

\item{axisTick}{A list. Default: \cr
\code{show=FALSE, inside=FALSE, length=5, lineStyle=list(color="#333", width=1)} \cr \cr
\code{lineStyle} accepts feature \code{color, width, type, shadowColor, shadowBlur,
shadowOffsetX, shadowOffsetY}}

\item{axisLabel}{A list controlling the axis labels. Default \code{show=TRUE,
rotate=0, margin=8, clickable=FALSE, formatter=NULL, textStyle=list(color="#333")} \cr \cr
\code{textStyle} accepts features \code{color, align, baseline, fontFamily, fontSize,
fontStyle, fontWeight}. \cr \cr
\strong{\code{formatter}}:
\describe{
 \item{sprintf/format string}{String to overide \code{axisLable$formatter}.
 It accepts \code{sprintf} (category and value) and \code{strptime} (time) formats.}
 \item{js mode}{a JS function/expression, which is default}
} \cr
\code{axisLabel=list(formatter="\%s cm")} is equal to \cr
\code{axisLabel=list(formatter=JS('function (value) {return value + "cm";}'))} or \cr
\code{axisLabel=list(formatter='{value} cm')}}

\item{splitLine}{A list controlling the split lines. Default \code{show=TRUE,
lineStyle=list(color=list("#ccc"), width=1, type="solid")} \cr \cr
\code{lineStyle} accepts features \code{color, width, type, shadowColor, shadowBlur,
shadowOffsetX, shadowOffsetY}.}

\item{splitArea}{A list controlling the split areas. Default \code{show=FALSE,
onGap=NULL, areaStyle= list(color= list("rgba(250,250,250,0.3)", "rgba(200,200,200,0.3)")
, type="default"} \cr \cr
\code{areaStyle} accepts features \code{color, type}.}

\item{data}{A character vector/list for axis of type 'category', to define the text
labels shown in this axis. Default NULL. You can even pass in a complicated list with
\code{textSytle} list: \cr
\code{list('Jan', 'Feb', 'Mar', \cr list(value='Apr', textStyle=list(color='red', ...)),
\cr 'May', ...)}}
}
\description{
When an echart object is generated, you can modify it by setting axis using
\code{\link{\%>\%}}. \cr \cr
You can use work functions \code{setXAxis}, \code{setYAxis}, \code{setX1Axis},
\code{setY1Axis}. \cr \cr
This function modified a few default options for the axis component in
ECharts:
\enumerate{
 \item \code{scale = TRUE} (was \code{FALSE} by default in ECharts);
 \item \code{axisLine$onZero = FALSE} (was \code{TRUE} in ECharts).
}
}
\examples{
\dontrun{
g <- echartR(iris, Sepal.Width, Petal.Width, series=Species)

# Change the style
g \%>\% setTheme('gray') \%>\% setXAxis(splitLine=list(show=FALSE)) \%>\%
      setYAxis(axisLine=list(lineStyle=list(width=0)))

# Dual-yAxis, series 1,2 on primary y-axis, series 3 on secondary y-axis
g \%>\% setYAxis(1:2, name="setosa/versicolor") \%>\%
      setY1Axis("virginica", name="virginica")
}
}
\references{
\url{http://echarts.baidu.com/echarts2/doc/option.html#tooltip-line1~xAxis-i}

\url{http://echarts.baidu.com/echarts2/doc/option.html#tooltip-line1~yAxis-i}
}
\seealso{
\code{\link{setXAxis}}, \code{\link{setYAxis}}, \code{\link{setX1Axis}},
\code{\link{setY1Axis}}
}

