% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{setDataRange}
\alias{setDataRange}
\title{Set \code{dataRange} Bar of Echarts}
\usage{
setDataRange(chart, show = TRUE, pos = 8, valueRange = NULL,
  splitNumber = 5, itemGap = 5, labels = NULL, calculable = FALSE,
  borderColor = "#ccc", borderWidth = 0, selectedMode = list(TRUE,
  "single", "multiple"), color = c("#1e90ff", "#f0ffff"), splitList = NULL,
  initialRange = NULL, ...)
}
\arguments{
\item{chart}{\code{echarts} object generated by \code{\link{echart}} or \code{\link{echartR}}}

\item{show}{logical. Show the dataRange control if TRUE. If you want to remove
dataRange from the echarts object, set it NULL.}

\item{pos}{the clock-position of dataRange, default 8. Refer to \code{\link{vecPos}}.
Or you can define the position vector \code{c(x, y, orient)} yourself.}

\item{valueRange}{The range of the dataRange bar in form of \code{c(min, max)}.
If NULL, echarts default will be used.}

\item{splitNumber}{How many discrete sections will the dataRange bar be divided into.
Default 5. Set it to 0 to set the bar continuous and calculable will be accordingly set TRUE.}

\item{itemGap}{The gap between itmes in pixels. Default 10px.}

\item{labels}{The labels to the ends the dataRange bar in form \code{c('high end',
'low end')}. Default NULL, the min, max values will be used.}

\item{calculable}{Logical. If echart calculable feature is open. Default FALSE. If set
calculable TRUE, splitNumber will be set 0.}

\item{borderColor}{The border color of the dataRange bar. Default '#333'.}

\item{borderWidth}{The border width of the dataRange bar. Default 0px (not shown).}

\item{selectedMode}{The mode of the dataRange bar, default TRUE. You can also
set it 'single' or 'multiple'.}

\item{color}{The hex vector of colors used for dataRange bar. Default c("#1e90ff", "#f0ffff").}

\item{splitList}{A list for user-defined value split in the form of
\code{list(list(start=m, end=n, label=x, color=hex), ...)}. If a valid splitList is set,
splitNumber will be disabled.}

\item{initialRange}{Initial selected value range in the form of \code{list(start=m, end=n)}}

\item{...}{Elipsis}
}
\value{
A modified echarts object
}
\description{
When an echart object is generated, you can modify it by setting dataRange using
\code{\link{\%>\%}}.
}
\examples{
\dontrun{
g <- iris \%>\% echartR(x=Sepal.Width, y=Petal.Width, series=Species, type='scatter')
g \%>\% setDataRange()

g1 <- iris \%>\% echartR(x=Sepal.Width, y=Petal.Width, z=Species, type='scatter')
g1 \%>\% setDataRange(valueRange=c(0, 2.5))
}
}
\references{
\url{http://echarts.baidu.com/echarts2/doc/option.html#title~dataRange}
}

