% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{setDataZoom}
\alias{setDataZoom}
\title{Set \code{dataZoom} Bar of Echarts}
\usage{
setDataZoom(chart, show = TRUE, pos = 6, range = NULL, width = 30,
  fill = "rgba(144,197,237,0.2)", handle = "rgba(70,130,180,0.8)",
  bgColor = "rgba(0,0,0,0)", dataBgColor = "#eee", showDetail = TRUE,
  realtime = FALSE, zoomLock = FALSE, ...)
}
\arguments{
\item{chart}{\code{echarts} object generated by \code{\link{echart}} or \code{\link{echartR}}}

\item{show}{logical. Show the dataZoom control if TRUE. If you want to remove
dataZoom from the echarts object, set it NULL.}

\item{pos}{the clock-position of dataZoom, refert to \code{\link{vecPos}}. You can
define a vector \code{c(x, y, orient)} yourself.}

\item{range}{A vector of \code{c(min, max)}. Cannot be out of the frame c(0, 100)}

\item{width}{The width of the dataZoom bar. Default 20px.}

\item{fill}{fillerColor of the dataZoom bar, in character \code{'rgba(red, green,
blue, alpha)'} format. Default 'rgba(144,197,237,0.2)' ("#90C5ED33").}

\item{handle}{handleColor of the dataZoom bar, in character \code{'rgba(red, green,
blue, alpha)'} format. Default 'rgba(70,130,180,0.8)' ("#4682B4CC").}

\item{bgColor}{background color. Default transparent ('rgba(0,0,0,0)')}

\item{dataBgColor}{background color of the data thumbnail (1st series). Default
'#eee'.}

\item{showDetail}{Logical, if show the details when zooming. Defaul TRUE.}

\item{realtime}{Logical, if realtime display the changes when zooming. Default FALSE.}

\item{zoomLock}{Logical, if the zoom range is locked. Deafult FALSE.}

\item{...}{Elipsis}
}
\value{
A modified echart object
}
\description{
When an echart object is generated, you can modify it by setting dataZoom using
\code{\link{\%>\%}}.
}
\examples{
\dontrun{
g <- iris \%>\% echartR(x=Sepal.Width, y=Petal.Width, series=Species, type='scatter')
g \%>\% setDataZoom(fill=rgba(c(col2rgb('gold'), 0.3)),
                  handle=rgba(c(col2rgb('gold'), 1)))

g1 <- iris \%>\% echartR(x=Sepal.Width, y=Petal.Width, z=Species, type='scatter')
g1 \%>\% setDataZoom(fill=rgba(c(col2rgb('lightgreen'), 0.2)),
                  handle=rgba(c(col2rgb('darkgreen'), 0.5)))
}
}
\references{
\url{http://echarts.baidu.com/echarts2/doc/option.html#title~dataZoom}
}

