% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{setLegend}
\alias{setLegend}
\title{Set \code{legend} of Echarts}
\usage{
setLegend(chart, show = TRUE, pos = 11, selected = NULL, itemGap = 5,
  borderColor = "#ccc", borderWidth = 0, textStyle = list(color = "auto"),
  formatter = NULL, overideData = NULL, ...)
}
\arguments{
\item{chart}{\code{echarts} object generated by \code{\link{echart}} or \code{\link{echartR}}}

\item{show}{Logical. Show the legend or not. Default TRUE. If you want to remove
legend from the echarts object, set it NULL.}

\item{pos}{Clock position of the legend. Default 11. Refer to \code{\link{vecPos}}.
Or you can define the position vector \code{c(x, y, orient)} yourself.}

\item{selected}{A vector of series names that are selected on load. If you assign 'none',
then none of the series will be selected in the beginning.}

\item{itemGap}{The gap between legend items. Default 5px.}

\item{borderColor}{The border color of the legend. Default '#ccc'.}

\item{borderWidth}{The border width of the legend. Default 0px (not shown).}

\item{textStyle}{A list of textStyle definition to decorate the text. E.g.,
\code{list(color='#444')} or \code{list(color='auto')}.}

\item{formatter}{A named formatter template or a string containing javascript codes.
E.g., \code{'{name}'}.}

\item{overideData}{A list of data to overide the legend text. E.g.,
\code{list(list(name='Series 1', icon='image://../asset/ico/favicon.png',
textStyle=list(color='#bbb')))}}

\item{...}{Elipsis.}
}
\value{
A modified echarts object.
}
\description{
When an echart object is generated, you can modify it by setting legend using
\code{\link{\%>\%}}.
You can modify the legend of the echarts object using this function.
}
\examples{
\\donrun{
# No timeline
g <- iris \%>\% echartR(x=Sepal.Width, y=Petal.Width, series=Species,
                      type='scatter')
g \%>\% setLegend(selected="versicolor")
g \%>\% setLegend(selected=levels(iris$Species)[1:2],
                textStyle=textStyle(fontFamily='Times New Roman', color='purple',
                               fontWeight='bold', fontSize=16))

# With Timeline
g1 <- iris \%>\% echartR(x=Sepal.Width, y=Petal.Width, z=Species, type='scatter')
g1 \%>\% setLegend(pos=12, selected='none',
                 textStyle=list(fontFamily='Courier New', fontSize=16))
}
}
\references{
\url{http://echarts.baidu.com/echarts2/doc/option.html#title~legend}
}

