% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{setPolar}
\alias{setPolar}
\title{Set \code{polar} of Echarts (For Radar Charts)}
\usage{
setPolar(chart, polarIndex = NULL, center = c("50\%", "50\%"),
  radius = "75\%", startAngle = 90, splitNumber = 5, boundaryGap = c(0,
  0), scale = FALSE, axisLine = NULL, axisLabel = NULL,
  splitLine = NULL, splitArea = NULL, type = c("polygon", "circle"),
  indicator = NULL, axisName = NULL, ...)
}
\arguments{
\item{chart}{\code{echarts} object generated by \code{\link{echart}} or \code{
\link{echartR}}}

\item{polarIndex}{Integer vector. The index of the polar systems you want to set.
Default NULL.}

\item{center}{Vector of the x, y position of the polar center. Could be numeric
or character (percent form) vectors of length 2. Default c('50\%', '50\%').}

\item{radius}{The radius of the polar system, could be numeric or character (percent form).
Default '75\%'.}

\item{startAngle}{Numeric (-180 ~ 180). The start angle. Default 90.}

\item{splitNumber}{Numeric. The number of sections to divide. Default 5.}

\item{boundaryGap}{Numeric vector of length 2. The gapping policy of the axis.
Default c(0, 0).}

\item{scale}{Logical. Whether to ignore zero and zoom toward the range of _min and _max.}

\item{axisLine}{List. Axis line styles. You can set its \code{show, onZero, lineStyle}
features. Default \code{list(show=TRUE)}.}

\item{axisLabel}{List. Axis label styles. You can set its \code{show, rotate, margin,
clickable, formatter, textStyle} features. Default \code{list(show=FALSE)}.}

\item{splitLine}{List. Split line styles. You can set its \code{show, lineStyle}
features. Default \code{list(show=TRUE)}.}

\item{splitArea}{List. Split area styles. You can set its \code{show, onGap, areaStyle}
features. Default \code{list(show=TRUE)}.}

\item{type}{Character, 'polygon' or 'circle'. The type of the polar shape.
Default 'polygon'.}

\item{indicator}{List. The radar indicator and labels. The basic structure is \code{
list(list(text='...', min=..., max=..., axisLabel=list(...)), list(text='...', min=..., max=...),
list(...), ...)}. Default is empty.}

\item{axisName}{List. The name of the axis. You can set its \code{show, formatter,
textStyle} features. Default \code{list(show=TRUE, formatter=NULL, textStyle=
list(color='#333'))}.}

\item{...}{Elipsis}
}
\value{
A modified echarts object
}
\description{
Set the \code{polar} coordinates of Echarts for radar charts.  \cr
When an echart object is generated, you can modify it by setting aesthetics using
\code{\link{\%>\%}}.
}
\examples{
\dontrun{
cars <- mtcars[c('Merc 450SE','Merc 450SL','Merc 450SLC'),
              c('mpg','disp','hp','qsec','wt','drat')]
cars$model <- rownames(cars)
cars <- data.table::melt(cars, id.vars='model')
names(cars) <- c('model', 'indicator', 'Parameter')
g <- echartr(cars, indicator, Parameter, model, type='radar') \%>\%
         setTitle('Merc 450SE  vs  450SL  vs  450SLC')
g \%>\% setPolar(c(1,3), type='circle') \%>\%
      setPolar(2, splitArea=list(show=FALSE)) \%>\%
      setPolar(3, axisName=list(textStyle=textStyle(color='red')))
}

}
\references{
\url{http://echarts.baidu.com/echarts2/doc/option.html#title~polar}
}

