% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{setSeries}
\alias{setSeries}
\title{Set `series` Element by Element}
\usage{
setSeries(chart, series = NULL, timeslots = NULL, ...)
}
\arguments{
\item{chart}{chart \code{echarts} object generated by \code{\link{echart}} or
\code{\link{echartR}}}

\item{series}{series, could be the name or the index of the series. When the string
is not in the existing series list, it will be treated as new}

\item{timeslots}{vector of timeslots, either name or index of timeline variable
\code{t}.}

\item{...}{The params to pass to the echarts object for modification. \cr
The generic params include:
\tabular{ll}{
 \code{type} \tab chart type. 'line', 'bar', 'scatter', 'k', 'pie', 'radar',
  'chord', 'force', 'map'... If NULL, the series will not display. \cr
 \code{name} \tab name of the series. \cr
 \code{tooltip} \tab \code{\link{setTooltip}} \cr
 \code{data} \tab the \code{data} part of \code{\link{echartR}} \cr
 \code{itemStyle} \tab \code{\link{itemStyle}} \cr
 \code{markPoint} \tab \code{\link{addMarkPoint}} \cr
 \code{markLine} \tab \code{\link{addMarkLine}} \cr
 \code{clickable} \tab whether the data graph is clickable, default TRUE. \cr
 \code{z} \tab level-2 layer control, default 2. \cr
 \code{zlevel} \tab level-1 layer control, default 0.
} \cr
The type-specific params include:
\describe{
 \item{line}{
   \tabular{ll}{
     \code{stack} \tab TRUE|FALSE \cr
     \code{xAxisIndex} \tab 0|1, default 0 (main axis) \cr
     \code{yAxisIndex} \tab 0|1, default 0 (main axis) \cr
     \code{symbol} \tab \code{\link{setSymbols}} \cr
     \code{symbolSize} \tab numeric or \code{\link{JS}} function \cr
     \code{symbolRotate} \tab -180 ~ 180 \cr
     \code{showAllSymbol} \tab TRUE|FALSE, default FALSE\cr
     \code{smooth} \tab TRUE|FALSE, default FALSE \cr
     \code{dataFilter} \tab only for line charts. 'nearest'|'min'|'max'|'average', default 'nearest'. \cr
     \code{legendHoverLink} \tab TRUE|FALSE, default TRUE.
   }}
 \item{bar}{
   \tabular{ll}{
     \code{stack} \tab TRUE|FALSE \cr
     \code{xAxisIndex} \tab 0|1, default 0 (main axis) \cr
     \code{yAxisIndex} \tab 0|1, default 0 (main axis) \cr
     \code{barGap} \tab percent or numeric, default '30\%'. \cr
     \code{barCategoryGap} \tab percent or numeric, default '20\%'. \cr
     \code{barMinHeight} \tab numeric, default 0. \cr
     \code{barWidth} \tab numeric, default automatic. \cr
     \code{barMaxWidth} \tab numeric, default automatic. \cr
     \code{legendHoverLink} \tab TRUE|FALSE, default TRUE.
     }}
 \item{scatter}{
   \tabular{ll}{
     \code{xAxisIndex} \tab 0|1, default 0 (main axis) \cr
     \code{yAxisIndex} \tab 0|1, default 0 (main axis) \cr
     \code{symbol} \tab \code{\link{setSymbols}} \cr
     \code{symbolSize} \tab numeric or \code{\link{JS}} function \cr
     \code{symbolRotate} \tab -180 ~ 180 \cr
     \code{large} \tab TRUE|FALSE, default FALSE. \cr
     \code{largeThreshold} \tab numeric, default 2000. \cr
     \code{legendHoverLink} \tab TRUE|FALSE, default TRUE.
     }}
 \item{pie}{
   \tabular{ll}{
     \code{legendHoverLink} \tab TRUE|FALSE, default TRUE. \cr
     \code{center} \tab percent or numeric, default c('50\%', '50\%') \cr
     \code{radius} \tab percent or numeric, default c(0, '75\%') \cr
     \code{startAngle} \tab numeric -360 ~ 360, default 90 for pie, 225 for gauge. \cr
     \code{minAngle} \tab numeric, default 0. \cr
     \code{clockWise} \tab TRUE|FALSE, default TRUE. \cr
     \code{roseType} \tab 'radius'|'area'. \cr
     \code{selectedOffset} \tab numeric, default 10. \cr
     \code{selectedMode} \tab NULL|'single'|'multiple', default NULL.
     }}
 \item{radar}{
   \tabular{ll}{
     \code{symbol} \tab \code{\link{setSymbols}} \cr
     \code{symbolSize} \tab numeric or \code{\link{JS}} function \cr
     \code{symbolRotate} \tab -180 ~ 180 \cr
     \code{legendHoverLink} \tab TRUE|FALSE, default TRUE. \cr
     \code{polarIndex} \tab numeric. Index of the polar system to host the series.
     }}
 \item{chord}{
   \tabular{ll}{
     \code{symbol} \tab \code{\link{setSymbols}} \cr
     \code{symbolSize} \tab numeric or \code{\link{JS}} function \cr
     \code{clockWise} \tab TRUE|FASLE, default FALSE. \cr
     \code{categories} \tab character vector \cr
     \code{links} \tab list of links \cr
     \code{matrix} \tab matrix data \cr
     \code{minRadius} \tab numeric, default 10. \cr
     \code{maxRadius} \tab numeric, default 20. \cr
     \code{ribbonType} \tab TRUE|FALSE, default TRUE. \cr
     \code{showScale} \tab TRUE|FALSE, default FALSE. \cr
     \code{showScaleText} \tab TRUE|FALSE, default FALSE \cr
     \code{padding} \tab numeric, default 2. \cr
     \code{sort} \tab 'none'|'ascending'|'descending', default 'none'. \cr
     \code{sortSub} \tab 'none'|'ascending'|'descending', default 'none'. \cr
     \code{nodes} \tab nodes list
     }}
 \item{force}{
   \tabular{ll}{
     \code{symbol} \tab \code{\link{setSymbols}} \cr
     \code{symbolSize} \tab numeric or \code{\link{JS}} function. \cr
     \code{large} \tab TRUE|FALSE, default FALSE. \cr
     \code{center} \tab percent or numeric, default c('50\%', '50\%'). \cr
     \code{roam} \tab TRUE|FALSE|'scale'|'move', default FALSE. \cr
     \code{categories} \tab character vector \cr
     \code{links} \tab links list \cr
     \code{matrix} \tab data matrix \cr
     \code{size} \tab percent or numeric, default '100\%' \cr
     \code{minRadius} \tab numeric, default 10. \cr
     \code{maxRadius} \tab numeric, default 20. \cr
     \code{linkSymbol} \tab vector of symbol characters, default c('none', 'arrow') \cr
     \code{linkSymbolSize} \tab numeric verctor, default c(10, 15) \cr
     \code{scaling} \tab numeric, default 1 \cr
     \code{gravity} \tab numeric, default 1 \cr
     \code{draggable} \tab TRUE|FALSE, default TRUE \cr
     \code{useWorker} \tab TRUE|FALSE, default FALSE \cr
     \code{steps} \tab numeric, default 1 \cr
     \code{nodes} \tab nodes list
     }}
 \item{map}{
   \tabular{ll}{
     \code{selectedMode} \tab NULL|'single'|'multiple', default NULL \cr
     \code{mapType} \tab 'china'|'world', default 'china' \cr
     \code{hoverable} \tab TRUE|FALSE, default TRUE \cr
     \code{dataRangeHoverLink} \tab TRUE|FALSE, default TRUE \cr
     \code{mapLocation} \tab list{x=.., y=.., width=.., height=..}, default
     list(x='center', y='center') \cr
     \code{mapValueCalculation} \tab 'sum'|'average', default 'sum' \cr
     \code{mapValuePrecision} \tab integer, default 0 \cr
     \code{showLegendSymbol} \tab TRUE|FALSE, default TRUE \cr
     \code{roam} \tab TRUE|FALSE|'scale'|'move', default FALSE \cr
     \code{scaleLimit} \tab numeric list(max=.., min=..), default NULL \cr
     \code{nameMap} \tab \code{\link{addNameMap}} \cr
     \code{textFixed} \tab list, e.g., list(China=c(10, -10)), default NULL \cr
     \code{geoCoord} \tab \code{\link{addGeoCoord}} \cr
     \code{heatmap} \tab \code{\link{addHeatMap}}
     }}
 \item{gauge}{
   \tabular{ll}{
     \code{legendHoverLink} \tab TRUE|FALSE, default TRUE. \cr
     \code{center} \tab percent or numeric vector, default c('50\%', '50\%') \cr
     \code{radius} \tab percent or numeric vector, default c(0, '75\%') \cr
     \code{startAngle} \tab numeric -360 ~ 360, default 225 \cr
     \code{endAngle} \tab numeric -360 ~ 360, default -45 \cr
     \code{min} \tab numeric, default 0 \cr
     \code{max} \tab numeric, default 100 \cr
     \code{splitNumber} \tab numeric, default 10 \cr
     \code{axisLine} \tab list, default \code{list(
      show=TRUE, lineStyle=list(
      color=list(list(0.2, '#228b22'), list(0.8, '#48b'), list(1, '#ff4500')),
      width=30))} \cr
     \code{axisTick} \tab list, default \code{list(
      show=TRUE, splitNumber=5, length=8, lineStyle=list(
      color='#eee', width=1, type='solid'))} \cr
     \code{axisLabel} \tab list, default \code{list(show=TRUE, formatter=NULL,
      textStyle=list(color='auto'))} \cr
     \code{splitLine} \tab list, default \code{list(show=TRUE,
      length=30, lineStyle=list(
      color='#eee', width=2, type='solid'))} \cr
     \code{pointer} \tab list, default \code{list(length='80\%', width=8,
      color='auto')} \cr
     \code{title} \tab list, default \code{list(show=TRUE, offsetCenter=list(0, '40\%'),
      textStyle=list(color='#333', fontSize=15))} \cr
     \code{detail} \tab list, default \code{list(show=TRUE, backgroundColor='rgba(0,0,0,0)',
      borderWidth=0, borderColor='#ccc', width=100, height=40, offsetCenter=list(0, '40\%'),
      formatter=NULL, textStyle=list(color='auto', fontSize=30))}
     }}
 \item{funnel}{
   \tabular{ll}{
     \code{legendHoverLink} \tab TRUE|FALSE, default TRUE. \cr
     \code{sort} \tab 'ascending'|'descending', default 'descending' \cr
     \code{min} \tab numeric, default 0 \cr
     \code{max} \tab numeric, default 100 \cr
     \code{x} \tab numeric, default 80 \cr
     \code{y} \tab numeric, default 60 \cr
     \code{x2} \tab numeric, default 80 \cr
     \code{y2} \tab numeric, default 60 \cr
     \code{width} \tab percent or numeric, default NULL \cr
     \code{height} \tab percent or numeric, default NULL \cr
     \code{funnelAlign} \tab 'left'|'right'|'center', default 'center' \cr
     \code{minSize} \tab percent, default '0\%' \cr
     \code{maxSize} \tab percent, default '100\%' \cr
     \code{gap} \tab numeric, default 0
     }}
 \item{eventRiver}{
   \tabular{ll}{
     \code{xAxisIndex} \tab 0|1, default 0 (main axis) \cr
     \code{legendHoverLink} \tab TRUE|FALSE, default TRUE. \cr
     \code{weight} \tab numeric, default 1
     }}
 \item{treemap}{
   \tabular{ll}{
     \code{center} \tab percent or numeric vector, default c('50\%', '50\%') \cr
     \code{size} \tab percent or numeric vector, default c('80\%', '80\%') \cr
     \code{root} \tab name of the root displayed currently
     }}
 \item{tree}{
   \tabular{ll}{
     \code{symbol} \tab \code{\link{setSymbols}} \cr
     \code{symbolSize} \tab numeric or \code{\link{JS}} function. \cr
     \code{roam} \tab TRUE|FALSE|'scale'|'move', default FALSE \cr
     \code{rootLocation} \tab list, could be list(x='center' | 'left' | 'right' | 'x\%' | {number},
     y='center' | 'top' | 'bottom' | 'y\%' | {number}) \cr
     \code{layerPadding} \tab numeric, default 100 \cr
     \code{nodePadding} \tab numeric, default 30 \cr
     \code{orient} \tab 'vertical' | 'horizontal' | 'radial', default 'vertical' \cr
     \code{direction} \tab "" or "inverse", default ""
     }}
 \item{wordCloud}{
   \tabular{ll}{
     \code{center} \tab percent or numeric vector, default c('50\%', '50\%') \cr
     \code{size} \tab percent or numeric vector, default c('80\%', '80\%') \cr
     \code{textRotation} \tab vector, -90 ~ 90, default c(0, 90) \cr
     \code{autoSize} \tab list, default \code{list(enable=TRUE, minSize=12)}
     }}
 \item{heatmap}{
   \tabular{ll}{
     \code{blurSize} \tab numeric, default 30 \cr
     \code{gradientColors} \tab \code{list(offset=.., color=..)} or color vector.
     default \code{list('blue', 'cyan', 'limegreen', 'yellow', 'red')} \cr
     \code{minAlpha} \tab numeric, default 0.05 \cr
     \code{valueScale} \tab numeric, default 1 \cr
     \code{opacity} \tab numeric, default 1
     }}
}}
}
\value{
A modified echarts object
}
\description{
The basic framework of making Echarts object is Step 1 \code{echart} => Step 2
\code{set[Widgets]} => Step 3 \code{set[Aesthetics]}. Step 1 makes the structure
of the Echarts object, step 2 creates/modifies its widgets, and step 3 modifies
the params of the widgets. But what if you want to modify the details inside
\code{series} list? -- You may want \code{setSeries}.\cr \cr
\code{setSeries} is a bit low-level, compared to other \code{set[Widgets]} functions.
It directly pass the param list on to an Echarts object and may lead to unexpected
results. So you are suggested to get more familiar with Echarts structure before
you decide to \code{setSeries}.
}
\examples{
\dontrun{
g = echartr(iris, Sepal.Width, Petal.Width, Species)
g \%>\% setSeries(series=3, symbol='emptyCircle')
}
}
\references{
\url{http://echarts.baidu.com/echarts2/doc/option.html#title~series-i}
}

