% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{setSymbols}
\alias{setSymbols}
\title{Set \code{symbolList} of Echarts}
\usage{
setSymbols(chart, symbols = NULL, ...)
}
\arguments{
\item{chart}{\code{echarts} object generated by \code{\link{echart}} or \code{\link{echartR}}}

\item{symbols}{A vector assigning symbols. You can use an array of symbols.
If the the symbols array is shorter than the levels of the series in length,
the last symbol will be used to extend the array. Default NULL.
\itemize{
 \item If you set symbolList NULL or
leave it unset, the function circulates the default symbol list of Echarts:
\code{c('circle', 'rectangle', 'triangle', 'diamond', 'emptyCircle', 'emptyRectangle',
'emptyTriangle', 'emptyDiamond')}.\cr
 \item You can also assign non-standard symbols,
such as \code{'heart', 'droplet', 'pin', 'arrow', 'star3', 'star4', 'star5', 'star6',
'star7', 'star8', 'star9'}, etc.\cr
 \item When assigned to \strong{'none', no symbol will be shown}, you can thus yield lines
without points on it.  \cr
 \item The elements not in the above eligible symbols list will be eliminated.
}}

\item{...}{Elipsis.}
}
\value{
A modified echarts object.
}
\description{
When an echart object is generated, you can modify it by setting symbolList using
\code{\link{\%>\%}}.
You can modify \code{symbolList} of the echarts object using this function. The symbolList
will be then mapped to data series in order.
}
\examples{
\dontrun{
g <- iris \%>\% echartR(x=Sepal.Width, y=Petal.Width, series=Species)
g \%>\% setSymbols('emptycircle')
g \%>\% setSymbols(c('heart', 'pin'))
}
}
\references{
\url{http://echarts.baidu.com/echarts2/doc/option.html#title~symbolList}
}

