% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{setTheme}
\alias{setTheme}
\title{Set Theme Or Misc Aesthetics (color, background, animation, border effects) of Echarts}
\usage{
setTheme(chart, theme = c("asis", "default", "macarons", "infographic",
  "blue", "dark", "gray", "green", "helianthus", "macarons2", "mint", "red",
  "roma", "sakura", "shine", "vintage"), palette = "asis", bgColor = NULL,
  renderAsImage = FALSE, calculable = FALSE, calculableColor = NULL,
  calculableHolderColor = NULL, animation = TRUE, animationEasing = NULL,
  animationDuration = NULL, width = NULL, height = NULL, ...)
}
\arguments{
\item{chart}{\code{echarts} object generated by \code{\link{echart}} or \code{
\link{echartR}}}

\item{theme}{\describe{
\item{Pre-installed themes}{\code{'default', 'macarons', 'infographic', 'blue',
'dark', 'gray', 'green', 'helianthus', 'macarons2', 'mint', 'red', 'roma',
'sakura', 'shine', 'vintage'}}
\item{User-defined themes}{E.g., \code{list(color=c('#ff7f50', '#87cefa',
'#da70d6','#32cd32','#6495ed'), \cr
title=list(x='left', y='top'), legend=list(orient='horizontal'))}}
}}

\item{palette}{name of the palette or a color vector. Default NULL to get echarts default.
It could be:
\describe{
 \item{\code{asis}}{keep the color palette applied in current echarts object}
 \item{Echarts theme palette}{"_default", "_macarons", "_infographic",
  "_blue", "_dark", "_gray", "_green", "_helianthus", "_macarons2", "_mint",
  "_red", "_roma", "_sakura", "_shine", "_vintage"}
 \item{\link{RColorBrewer} palettes}{Including 'BrBG', 'PiYG', 'PRGn', 'PuOr', 'RdBu',
 'RdGy', 'RdYlBu', 'RdYlGn', 'Spectral', 'Accent', 'Dark2', 'Paired', 'Pastel1',
 'Pastel2', 'Set1', 'Set2', 'Set3', 'Blues', 'BuGn', 'BuPu', 'GnBu', 'Greens',
 'Greys', 'Oranges', 'OrRd', 'PuBu', 'PuBuGn', 'PuRd', 'Purples', 'RdPu', 'Reds',
 'YlGn', 'YlGnBu', 'YlOrBr', 'YlOrRd'}
 \item{\link{ggthemes} palettes}{'calc', 'economist', 'economist_white', 'economist_stata',
 'excel', 'exel_fill', 'excel_line', 'excel_new', 'few', 'fivethirtyeight', '538', 'manyeyes',
 'gdocs', 'pander', 'tableau', 'stata', 'stata1','stata1r','statamono', 'ptol',
 'tableau20', 'tableau10medium', 'tableaugray', 'tableauprgy', 'tableaublrd',
 'tableaugnor', 'tableaucyclic', 'tableau10light', 'tableaublrd12', 'tableauprgy12',
 'tableaugnor12','hc','darkunica', 'solarized','solarized_red', 'solarized_yellow',
 'solarized_orange', 'solarized_magenta', 'solarized_violet', 'solarized_blue',
 'solarized_cyan', 'solarized_green', 'wsj', 'wsj_rgby', 'wsj_red_green',
 'wsj_black_green', 'wsj_dem_rep', 'colorblind', 'trafficlight'}
 \item{Aetna official palettes}{Including 'aetnagreen', 'aetnablue', 'aetnaviolet',
 'aetnaorange', 'aetnateal', 'aetnacranberry'}
 \item{Other palettes}{'rainbow', 'terrain', 'topo', 'heat', 'cm'}
}
\strong{Usage:} \cr
\itemize{
 \item If the value is not set, and the function defaults will be loaded \cr
 \item Set palette=NULL to use Echarts defaults \cr
 \item Set palette=palette name to assign any palette listed above \cr
 \item Set palette=\code{palette name(number)} to restrict number of colors within the
 palette (e.g., \code{palette='calc(3)'} picks 3 colors out of 'calc' \strong{RANDOMLY}) \cr
 \item Set palette=\code{c(color1, color2, color3, ...)} to define a palette vector,
 made of which either color names or Hex expressions. Use \code{\link{colors}()} to check
 available color names and check the effects using \code{demo(colors)}.
}}

\item{bgColor}{Color name/value of the background. Default is transparent
(\code{'rgba(0,0,0,0)'})}

\item{renderAsImage}{Logical. If TRUE, the interactive effects are disabled. Default FALSE.}

\item{calculable}{Logical. If TRUE, the chart is re-calculated after drag. Default FALSE.}

\item{calculableColor}{The border color of the tooltip during \code{calculable} effect.
Default 'rgba(255,165,0,0.6)'.}

\item{calculableHolderColor}{The color of \code{calculableHolder}. Default '#ccc'.}

\item{animation}{Logical. If TRUE, the animation is on at initiation. Default TRUE.
For IE8, it is recommended to set FALSE.}

\item{animationEasing}{The slight moving effect of major elements. Default 'ExponentialOut'.
Other choices: 'Linear', 'QuadraticIn', 'QuadraticOut', 'QuadraticInOut', 'CubicIn',
'CubicOut', 'CubicInOut', 'QuarticIn', 'QuarticOut', 'QuarticInOut', 'QuinticIn',
'QuinticOut', 'QuinticInOut', 'SinusoidalIn', 'SinusoidalOut', 'SinusoidalInOut',
'ExponentialIn', 'ExponentialInOut', 'CircularIn', 'CircularOut', 'CircularInOut',
'ElasticIn', 'ElasticOut', 'ElasticInOut', 'BackIn', 'BackOut', 'BackInOut',
'BounceIn', 'BounceOut', 'BounceInOut'}

\item{animationDuration}{The duration into animation. Default 2000 (ms).}

\item{width}{Width (px) of the whole chart.}

\item{height}{Height (px) of the whole chart.}

\item{...}{Elipsis}
}
\value{
A modified echarts object
}
\description{
Set the entire theme of Echarts. You can use a pre-installed theme or define
you own themes. \cr \cr
You can also set aesthetics of Echarts separately, including \code{color, bgColor,
animation, calculable, borderColor, borderWidth}.\cr \cr
When an echart object is generated, you can modify it by setting aesthetics using
\code{\link{\%>\%}}.
}
\section{Self-defined Themes (UDT)}{

You can design your prefered themes using
\url{http://echarts.baidu.com/echarts2/doc/example/themeDesigner.html}.\cr
\cr The UDTs in \code{echarts} are JS objects. You can write it in a list form
and \code{setTheme} will parse it into JSON string for process. \cr \cr
A typical theme JS looks like: \cr
\code{{\cr
 color: ['#ff7f50','#87cefa','#da70d6','#32cd32','#6495ed'], \cr
 title: {x: 'left', y: 'top'}, \cr
 legend: {orient: 'horizontal'} \cr
}} \cr \cr
You should write in the format below: \cr
\code{list(color=c('#ff7f50','#87cefa','#da70d6','#32cd32','#6495ed'), \cr
 title=list(x='left', y='top'), legend=list(orient='horizontal'))}
}
\examples{
\dontrun{
g <- echartR(iris, Sepal.Width, Petal.Width, series=Species)

# Set themes
## pre-installed themes
g \%>\% setTheme("infographic")
g \%>\% setTheme("blue")
g \%>\% setTheme("dark")
g \%>\% setTheme("gray")
g \%>\% setTheme("green")
g \%>\% setTheme("helianthus")
g \%>\% setTheme("macarons2")
g \%>\% setTheme("mint")
g \%>\% setTheme("red")
g \%>\% setTheme("sakura")
g \%>\% setTheme("shine")
g \%>\% setTheme("vintage")

## self-defined themes
theme <- list(color=c("#7AC143", "#7D3F98", "#F47721", "#D20962", "#00A78E",
              "#00BCE4", "#B8D936", "#EE3D94"), backgroundColor="#fef8ef")
g \%>\% setTheme(theme)

## Misc aethetics
g <- echartR(iris, x=Sepal.Width, y=Petal.Width, series=Species, type='scatter')
g \%>\% setTheme(palette='hc')
g \%>\% setTheme(palette=c('red', 'gold', 'deepskyblue'), 'gray95',
        animationHoldColor='red', animationEasing='CircularOut',
        animationDuration=10000)
}
}
\references{
\code{theme} \cr
\url{http://echarts.baidu.com/echarts2/doc/example/theme.html} \cr \cr
\code{backgroundColor} \cr
\url{http://echarts.baidu.com/echarts2/doc/option.html#title~backgroundColor} \cr \cr
\code{colors} \cr
\url{http://echarts.baidu.com/echarts2/doc/option.html#title~color} \cr
\url{http://colorbrewer2.org} \cr \cr
\code{other features} \cr
\url{http://echarts.baidu.com/echarts2/doc/option.html#title~renderAsImage} \cr
\url{http://echarts.baidu.com/echarts2/doc/option.html#title~calculable} \cr
\url{http://echarts.baidu.com/echarts2/doc/option.html#title~animation} \cr
\url{http://echarts.baidu.com/echarts2/doc/option.html#title~calculableColor} \cr
\url{http://echarts.baidu.com/echarts2/doc/option.html#title~calculableHolderColor}
}
\seealso{
\code{\link{brewer.pal}}, \code{\link{colors}}
}

