% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{setTitle}
\alias{setTitle}
\title{Set \code{title} (And Subtitle) of Echarts}
\usage{
setTitle(chart, title = NULL, subtitle = NULL, link = NULL,
  sublink = NULL, pos = 6, bgColor = NULL, borderColor = NULL,
  borderWidth = NULL, textStyle = NULL, subtextStyle = NULL,
  show = TRUE, ...)
}
\arguments{
\item{chart}{\code{echarts} object generated by \code{\link{echart}} or \code{\link{echartR}}}

\item{title}{text of the title. If written in markdown format \code{[caption](url)},
then \code{caption} is passed to title, \code{url} is passed to link. If the length
of the title vector equals to the length of timeline slices, the title vector will
be used as slice-specific user-defined title.}

\item{subtitle}{text of the subtitle. If written in markdown format \code{[caption](url)},
then \code{caption} is passed to title, \code{url} is passed to link. If the length
of the subtitle vector equals to the length of timeline slices, the subtitle vector will
be used as slice-specific user-defined title.}

\item{link}{link of the title}

\item{sublink}{link of the subtitle}

\item{pos}{the clock-position of title (and subtitle), refer to \code{\link{vecPos}}. Or
define a vector \code{c(x, y, orient)} yourself.}

\item{bgColor}{background color of title. Default 'rgba(0,0,0,0)' (transparent)}

\item{borderColor}{border color of the title. Default '#ccc'.}

\item{borderWidth}{border width of the title. Default 0px (not shown).}

\item{textStyle}{You can place self-defined textStyle list of the title here. E.g., \code{list(
fontFamily='Arial|Verdana|sans-serif', fontSize=20, fontWeight='normal|bold|bolder|lighter|<numbers>',
fontStyle='normal|italic|oblique', color='red')}}

\item{subtextStyle}{You can place self-defined textStyle of the subtitle here.}

\item{show}{Logical. Whether to show the title. If you want to remove title from
the echarts object, set it NULL.}

\item{...}{elipsis}
}
\value{
A modified echart object
}
\description{
When an echart object is generated, you can modify it by setting title and
subtitles using \code{\link{\%>\%}}.
}
\details{
In echarts with timeline, you can set \code{title, subtitle, link, sublink, bgColor,
borderColor, borderWidth} as vectors of the same length as \code{z}. For \code{textSytle,
subtextStyle}, you need to pack the user-defined lists in a list.
}
\examples{
\dontrun{
## simple echarts scatter plot
g <- iris \%>\% echartR(x=Sepal.Width, y=Petal.Width, series=Species,
                      type='scatter')

## set simple titles
g \%>\% setTitle(title='Iris data set', subtitle='by: R. A. Fisher')

## set titles with links and textStyles
g \%>\% setTitle(
  '[Iris data set](https://en.wikipedia.org/wiki/Iris_flower_data_set)',
  '[R. A. Fisher](https://en.wikipedia.org/wiki/Ronald_Fisher)',
  textStyle=textStyle(fontFamily='Courier New', fontSize=24, color='gold'),
  subtextStyle=textStyle(color='silver'), bgColor='lightgreen')


## echarts with timeline
g1 <- iris \%>\% echartR(x=Sepal.Width, y=Petal.Width, z=Species, type='scatter')

## simple titles/subtitles
g1 \%>\% setTitle(
    '[Iris data set](https://en.wikipedia.org/wiki/Iris_flower_data_set)',
    '[R. A. Fisher](https://en.wikipedia.org/wiki/Ronald_Fisher)')

## user-defined titles/subtitles
titles <- c(
    '[Iris setosa](https://en.wikipedia.org/wiki/Iris_setosa)',
    '[Iris versicolor](https://en.wikipedia.org/wiki/Iris_versicolor)',
    '[Iris virginica](https://en.wikipedia.org/wiki/Iris_virginica)')

g1 \%>\%
setTitle(
   titles,
   textStyle=list(
      textStyle(fontFamily='Impact', color='red', fontStyle='normal'),
      textStyle(fontFamily='Times New Roman', color='green'),
      textStyle(fontFamily='Calibri', color='blue', fontStyle='oblique')
   ),
   bgColor='lightyellow')
## textStyle is a list length 3, mapping 3 levels for timeline
}
}
\references{
\url{http://echarts.baidu.com/echarts2/doc/option.html#title~title}
}

