% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{setToolbox}
\alias{setToolbox}
\title{Set \code{toolbox} of Echarts}
\usage{
setToolbox(chart, show = TRUE, language = "cn", controls = c("mark",
  "dataZoom", "dataView", "magicType", "restore", "saveAsImage"), pos = 1,
  bgColor = "rgba(0,0,0,0)", borderColor = "#ccc", borderWidth = 0,
  padding = 5, itemGap = 10, itemSize = 16, color = c("#1e90ff",
  "#22bb22", "#4b0082", "#d2691e"), disableColor = "#ddd",
  effectiveColor = "red", showTitle = TRUE, textStyle = NULL, ...)
}
\arguments{
\item{chart}{\code{echarts} object generated by \code{\link{echart}} or \code{\link{echartR}}}

\item{show}{logical. Show the toolbox if TRUE. If you want to remove toolbox from
the echarts object, set it NULL.}

\item{language}{'cn' or 'en', the language of the toolbox tooltips.}

\item{controls}{which widgets to show. Default \code{'mark', 'dataZoom', 'dataView', 'magicType',
'restore', 'saveAsImage'}.}

\item{pos}{the clock-position of toolbox, refer to \code{\link{vecPos}}. Or you can
define a vector \code{c(x, y, orient)} yourself.#'}

\item{bgColor}{background color, default transparent ('rgba(0,0,0,0)').}

\item{borderColor}{border color, default '#ccc'.}

\item{borderWidth}{border width, default 0px (not shown).}

\item{padding}{Padding of the toolbox. Default 5px. Could also be a list assigning
padding of top, right, bottom and left.}

\item{itemGap}{Gap between the items. Default 10px.}

\item{itemSize}{Size of the items. Default 16px.}

\item{color}{Colors of the toolbox widgets (applied in loops). Default
c("#1e90ff", "#22bb22", "#4b0082", "#d2691e").}

\item{disableColor}{Color for disabled widgets. Default '#ddd'.}

\item{effectiveColor}{Color for widgets be triggered. Default 'red'.}

\item{showTitle}{Logical, if widgets title are shown. Default TRUE.}

\item{textStyle}{A list of the text style of the widgets. Default \code{
list(fontFamily=c('Arial, Verdana, sans-serif'), fontSize=12, fontStyle='normal',
fontWeight='normal')}}

\item{...}{elipsis}
}
\value{
A modified echart object
}
\description{
When an echart object is generated, you can modify it by setting toolbox using
\code{\link{\%>\%}}.
}
\examples{
\dontrun{
g <- iris \%>\% echartR(x=Sepal.Width, y=Petal.Width, series=Species, type='scatter')
g \%>\% setToolbox(TRUE, 'en')
}
}
\references{
\url{http://echarts.baidu.com/echarts2/doc/option.html#title~toolbox}
}

