% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{setTooltip}
\alias{setTT}
\alias{setTooltip}
\title{Set \code{tooltip} of Echarts}
\usage{
setTooltip(chart, series = NULL, timeslots = NULL, trigger = NULL,
  formatter = NULL, islandFormatter = "{a} < br/>{b} : {c}",
  position = NULL, enterable = FALSE, axisPointer = NULL,
  textStyle = NULL, showDelay = 20, hideDelay = 100,
  transitionDuration = 0.4, bgColor = "rgba(0,0,0,0.7)",
  borderColor = "#333", borderWidth = 0, borderRadius = 4, show = TRUE,
  ...)

setTT(chart, series = NULL, timeslots = NULL, trigger = NULL,
  formatter = NULL, islandFormatter = "{a} < br/>{b} : {c}",
  position = NULL, enterable = FALSE, axisPointer = NULL,
  textStyle = NULL, showDelay = 20, hideDelay = 100,
  transitionDuration = 0.4, bgColor = "rgba(0,0,0,0.7)",
  borderColor = "#333", borderWidth = 0, borderRadius = 4, show = TRUE,
  ...)
}
\arguments{
\item{chart}{\code{echarts} object generated by \code{\link{echart}} or \code{\link{echartR}}}

\item{series}{A vector of series indices or names. e.g., \code{c('setosa', 'virginica')}
or \code{1:2}}

\item{timeslots}{A vector of time slices indices or names, e.g., \code{c(1990, 1992)}
or \code{c(1,3)}. You can also use \code{t} as a short form of \code{timeslots}.}

\item{trigger}{Type of trigger, \code{'item'}, or \code{'axis'}.}

\item{formatter}{The format of the tooltip content.
\describe{
 \item{string(template)}{
  \itemize{
  \item \{a\} | \{a0\}
  \item \{b\} | \{b0\}
  \item \{c\} | \{c0\}
  \item \{d\} | \{d0\} (not applicable for some types)
  \item \{e\} | \{e0\} (not applicable for some types)
  }}
 \item{function}{the JS list is in the form \code{[params, ticket, callback]}.}
}
\strong{Meanings of \{a\}, \{b\}, \{c\}, \{d\}...}
\tabular{ll}{
 line, bar, k \tab \code{a} (series name), \code{b} (category name), \code{c} (value) \cr
 scatter \tab \code{a} (series name), \code{b} (data name), \code{c} (value array) \cr
 map \tab\code{a}(series name), \code{b} (area name), \code{c} (combined value) \cr
 pie, radar, gauge, funnel \tab \code{a} (series names), \code{b}(data item name),
   \code{c} (value), \code{d} (pie:percent|radar:indicator) \cr
 force, chord \tab \itemize{
   \item nodes: \code{a} (series name), \code{b} (node name), \code{c} (node value),
     \code{d} (node type index);
   \item links: \code{a} (series name), \code{b} (link name), \code{c} (link value),
     \code{d} (big node name/index), \code{e} (small node/index)
   }
}
\strong{JS Function Param Template...}
\tabular{ll}{
 \code{seriesIndex} \tab 0, 1, 2, ... \cr
 \code{seriesName} \tab 'Monday', 'Tuesday', ... \cr
 \code{name} \tab 'day1', 'day2', ... \cr
 \code{dataIndex} \tab 0, 1, 2, ... \cr
 \code{data} \tab data \cr
 \code{value} \tab value \cr
 \code{percent} \tab special  //pie \cr
 \code{indicator} \tab special  //radar, force, chord \cr
 \code{value2} \tab special2  //force, chord \cr
 \code{indicator2} \tab special2  //force, chord
}}

\item{islandFormatter}{Formatter of data island for calcualable effect. Can be
string (default \code{'{a} <br/>{b} : {c}'}) or JS function.}

\item{position}{Can be fixed position array \code{c(x, y)} or a JS function, e.g.,
\code{JS('function([x, y]) {return [newX, newY]}')}. Default NULL.}

\item{enterable}{If users can click into the tooltip for interacions. Default FALSE.}

\item{axisPointer}{The pointer formatter of axis. Default is a list \code{
list(type = "line", \cr
     lineStyle = list(color = "#48b", width = 2, type = "solid"), \cr
     crossStyle = list(color = "#1e90ff", width = 1, type = "dashed"), \cr
     shadowStyle = list(color = "rgba(150,150,150,0.3)", width = "auto", type = "default") \cr
)}.}

\item{textStyle}{text style of the tooltip. In a list form. Default \code{list(color
="#fff")}. The list coud contain elements of \code{color, decoration, fontSize,
fontFamily, fontStyle, fontWeight, align, baseline}.}

\item{showDelay}{Delayed time at show (ms). Default 20ms.}

\item{hideDelay}{Delayed time at hide (ms). Default 100ms.}

\item{transitionDuration}{The time spent at animation exchange. Default 0.4. Set
if 0 if you want real-time interaction.}

\item{bgColor}{Background color of tooltips. Default 'rgba(0,0,0,0.7)' (
semi-transparent dark gray).}

\item{borderColor}{Borderline color of the tooltips. Default '#333'.}

\item{borderWidth}{Border width of the tooltips. Default 0 (not shown).}

\item{borderRadius}{Border radius of the tooltips. Default 4px.}

\item{show}{Logical. If the tooltips are shown. Default TRUE.}

\item{...}{Elipsis}
}
\value{
A modified echarts object
}
\description{
Set tooltip of Echarts, at various levels (entire chart, specific series) with
various formats. \cr
When an echart object is generated, you can modify it by setting tooltip using
\code{\link{\%>\%}}.
}
\note{
Note that the \code{tooltip} feature is inheritable in terms of timeline.
\code{setTooltip} automatically breaks the inheritability by resetting tooltip
formats in the timeslots following the timeslots whose tooltip format are changed.
.
}
\examples{
\dontrun{
g = echartR(iris, Sepal.Width, Petal.Width, series=Species)

## global tooltip
g \%>\% setTooltip(trigger='item', bgColor='rgba(0,0,200,0.7)')

## series-specific tooltip
bg = sapply(c('orange', 'deepskyblue', 'violet'), col2rgb)
bg = rbind(bg, 0.7)  # extend the color matrix with alpha 0.7
bgCol = unname(apply(bg, 2, rgba))  # get rgba colors with alpha
g \%>\% setTooltip(series='setosa', bgColor=bgCol[1]) \%>\%
      setTooltip(series=2, bgColor=bgCol[2]) \%>\%
      setTooltip(series=3, bgColor=bgCol[3])

## series-and-timeline-specific tooltip
bg = sapply(c('orange', 'deepskyblue', 'violet'), col2rgb)
bg = rbind(bg, 0.7)  # extend the color matrix with alpha 0.7
bgCol = unname(apply(bg, 2, rgba))  # get rgba colors with alpha
iris$tag = 1 + as.integer(row.names(iris)) \%\% 3
iris = iris[order(iris$tag),]
g1 = echartR(iris, Sepal.Width, Petal.Width, series=Species, t=tag)
g1 \%>\% setTooltip(series='setosa', bgColor=bgCol[1]) \%>\%
       setTooltip(series=2, bgColor=bgCol[2]) \%>\%
       setTooltip(series=3, bgColor=bgCol[3]) \%>\%
       setTooltip(t=1, borderColor='red', borderWidth=3) \%>\%
       setTooltip(t=2, borderColor='gold', borderWidth=3) \%>\%
       setTooltip(t=3, borderColor='green', borderWidth=3)

g1 \%>\% setTooltip(series=2, t=2, bgColor='blue')
# tooltip format of series 2 in the 2nd timeslot is changed to bg blue,
# while it in the 3rd timeslot is reset to default
}
}
\references{
\url{http://echarts.baidu.com/echarts2/doc/option.html#title~tooltip}
}

